Change implementation of point_inside functions
This mitigates an issue if a point of the cell lies on the boundart line
as this would cause two neighboring cells being concidered boundary
cells. By moving to a implementation using less than instead of less or
equal only the cell where it is the outermost point is concidered a
boundary cell which seems like a sensibal way of handling this issue.